+++ src/select.js
@@ -249,7 +249,7 @@
         }
 
         if ( callCustomTrigger ) {
-            fireCustomTrigger( controlElements.bindingOptions.onSelectedItemsChanged );
+            fireCustomTrigger( controlElements.bindingOptions.onSelectedItemsChanged, getValuesSelected( controlElements ) );
         }
     }
 
@@ -293,7 +293,7 @@
                 controlElements.dropDown.style.display = "block";
 
                 renderDropDownItems( controlElements );
-                renderSelectedItems( controlElements );
+                renderSelectedItems( controlElements, false );
 
             }, controlElements.bindingOptions.dropDownShowDelay );
 
@@ -306,12 +306,28 @@
         if ( controlElements.dropDown !== null && controlElements.dropDown.style.display !== "none" ) {
             controlElements.dropDown.style.display = "none";
 
-            renderSelectedItems( controlElements );
+            renderSelectedItems( controlElements, false );
         }
     }
 
     function isDropDownMenuVisible( controlElements ) {
         return controlElements.dropDown !== null && controlElements.dropDown.style.display === "block";
+    }
+
+    function getValuesSelected( controlElements ) {
+        var options = controlElements.select.options,
+            optionsLength = options.length,
+            optionValuesSelected = [];
+
+        for ( var optionIndex = 0; optionIndex < optionsLength; optionIndex++ ) {
+            var option = options[ optionIndex ];
+
+            if ( option.selected ) {
+                optionValuesSelected.push( option.value );
+            }
+        }
+
+        return optionValuesSelected;
     }